refactor(runtime): extract runtime input drain
